+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/AgentSdJwtTest.kt
@@ -86,27 +86,30 @@ val AgentSdJwtTest by matrixSuite {
             val holderKeyMaterial = holderKeyMaterial
             val statusListIssuer = statusListIssuer
             val verifierId = "urn:${uuid4()}"
-            val verifier = VerifierAgent(
-                identifier = verifierId,
-                validatorSdJwt = validator,
+            val verifier = NonceChallengeVerifier(
+                verifierId = verifierId,
+                verifier = VerifierAgent(
+                    identifier = verifierId,
+                    validatorSdJwt = validator,
+                ),
             )
-            val challenge = uuid4().toString()
         }
     } }) - {
 
         "keyBindingJws contains more JWK attributes, still verifies" {
+            val request = it.verifier.createPresentationRequest()
             val credential = it.holderCredentialStore.getCredentials().getOrThrow()
                 .filterIsInstance<SubjectCredentialStore.StoreEntry.SdJwt>().first()
             val sdJwt = createSdJwtPresentation(
                 signKeyBindingJws = SignJwt(it.holderKeyMaterial, { header, _ ->
                     header.copy(keyId = "definitely not matching")
                 }),
-                audienceId = it.verifierId,
-                challenge = it.challenge,
+                audienceId = request.audience,
+                challenge = request.nonce,
                 validSdJwtCredential = credential,
                 claimName = CLAIM_GIVEN_NAME
             )
-            it.verifier.verifyPresentationSdJwt(sdJwt.sdJwt, it.challenge).getOrThrow()
+            it.verifier.verifyPresentationSdJwt(sdJwt.sdJwt).getOrThrow()
                 .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>().apply {
                     reconstructedJsonObject.keys shouldContain CLAIM_GIVEN_NAME
                     freshnessSummary.tokenStatusValidationResult
@@ -115,15 +118,16 @@ val AgentSdJwtTest by matrixSuite {
         }
 
         "presex: simple walk-through success" {
+            val request = it.verifier.createPresentationRequest()
             val presentationParameters = it.holder.createPresentation(
-                request = PresentationRequestParameters(nonce = it.challenge, audience = it.verifierId),
+                request = request,
                 credentialPresentation = buildPresentationDefinition(CLAIM_GIVEN_NAME, CLAIM_DATE_OF_BIRTH)
             ).getOrThrow().shouldBeInstanceOf<PresentationResponseParameters.PresentationExchangeParameters>()
 
             val vp = presentationParameters.presentationResults.firstOrNull()
                 .shouldBeInstanceOf<CreatePresentationResult.SdJwt>()
 
-            it.verifier.verifyPresentationSdJwt(vp.sdJwt, it.challenge).getOrThrow()
+            it.verifier.verifyPresentationSdJwt(vp.sdJwt).getOrThrow()
                 .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>().apply {
                     reconstructedJsonObject[CLAIM_GIVEN_NAME]?.jsonPrimitive?.content shouldBe "Susanne"
                     reconstructedJsonObject[CLAIM_DATE_OF_BIRTH]?.jsonPrimitive?.content shouldBe "1990-01-01"
@@ -133,27 +137,28 @@ val AgentSdJwtTest by matrixSuite {
         }
 
         "presex: wrong key binding jwt" {
+            val request = it.verifier.createPresentationRequest()
             val presentationParameters = it.holder.createPresentation(
-                request = PresentationRequestParameters(nonce = it.challenge, audience = it.verifierId),
+                request = request,
                 credentialPresentation = buildPresentationDefinition(CLAIM_GIVEN_NAME)
             ).getOrThrow().shouldBeInstanceOf<PresentationResponseParameters.PresentationExchangeParameters>()
 
             val vp = presentationParameters.presentationResults.firstOrNull()
                 .shouldBeInstanceOf<CreatePresentationResult.SdJwt>()
             // replace key binding of original vp.sdJwt (i.e. the part after the last `~`)
-            val freshKbJwt = createFreshSdJwtKeyBinding(it.challenge, it.verifierId)
+            val freshKbJwt = createFreshSdJwtKeyBinding(request.nonce, request.audience)
             val malformedVpSdJwt = vp.serialized.replaceAfterLast("~", freshKbJwt.substringAfterLast("~"))
 
             shouldThrowAny {
                 it.verifier.verifyPresentationSdJwt(
                     SdJwtSigned.parseCatching(malformedVpSdJwt).getOrThrow(),
-                    it.challenge
                 ).getOrThrow()
             }
         }
 
         "presex: wrong challenge in key binding jwt" {
-            val malformedChallenge = it.challenge.reversed()
+            val request = it.verifier.createPresentationRequest()
+            val malformedChallenge = request.nonce.reversed()
             val presentationParameters = it.holder.createPresentation(
                 request = PresentationRequestParameters(malformedChallenge, it.verifierId),
                 credentialPresentation = buildPresentationDefinition(CLAIM_GIVEN_NAME)
@@ -163,13 +168,14 @@ val AgentSdJwtTest by matrixSuite {
                 .shouldBeInstanceOf<CreatePresentationResult.SdJwt>()
 
             shouldThrowAny {
-                it.verifier.verifyPresentationSdJwt(vp.sdJwt, it.challenge).getOrThrow()
+                it.verifier.verifyPresentationSdJwt(vp.sdJwt).getOrThrow()
             }
         }
 
         "presex: revoked sd jwt" {
+            val request = it.verifier.createPresentationRequest()
             val presentationParameters = it.holder.createPresentation(
-                request = PresentationRequestParameters(nonce = it.challenge, audience = it.verifierId),
+                request = request,
                 credentialPresentation = buildPresentationDefinition(CLAIM_GIVEN_NAME)
             ).getOrThrow().shouldBeInstanceOf<PresentationResponseParameters.PresentationExchangeParameters>()
 
@@ -183,15 +189,16 @@ val AgentSdJwtTest by matrixSuite {
                         storeEntry.sdJwt.statusElement.shouldBeInstanceOf<StatusListInfo>().index
                     ) shouldBe true
                 }
-            it.verifier.verifyPresentationSdJwt(vp.sdJwt, it.challenge).getOrThrow()
+            it.verifier.verifyPresentationSdJwt(vp.sdJwt).getOrThrow()
                 .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
                 .freshnessSummary.tokenStatusValidationResult
                 .shouldBeInstanceOf<TokenStatusValidationResult.Invalid>()
         }
 
         "dcql: simple walk-through success" {
+            val request = it.verifier.createPresentationRequest()
             val presentationParameters = it.holder.createDefaultPresentation(
-                request = PresentationRequestParameters(nonce = it.challenge, audience = it.verifierId),
+                request = request,
                 credentialPresentationRequest = CredentialPresentationRequest.DCQLRequest(
                     buildDCQLQuery(
                         DCQLJsonClaimsQuery(
@@ -207,7 +214,7 @@ val AgentSdJwtTest by matrixSuite {
             val vp = presentationParameters.verifiablePresentations.values.flatten().firstOrNull()
                 .shouldBeInstanceOf<CreatePresentationResult.SdJwt>()
 
-            it.verifier.verifyPresentationSdJwt(vp.sdJwt, it.challenge).getOrThrow()
+            it.verifier.verifyPresentationSdJwt(vp.sdJwt).getOrThrow()
                 .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>().apply {
                     reconstructedJsonObject[CLAIM_GIVEN_NAME]?.jsonPrimitive?.content shouldBe "Susanne"
                     reconstructedJsonObject[CLAIM_DATE_OF_BIRTH]?.jsonPrimitive?.content shouldBe "1990-01-01"
@@ -217,8 +224,9 @@ val AgentSdJwtTest by matrixSuite {
         }
 
         "dcql: wrong key binding jwt" {
+            val request = it.verifier.createPresentationRequest()
             val presentationParameters = it.holder.createDefaultPresentation(
-                request = PresentationRequestParameters(nonce = it.challenge, audience = it.verifierId),
+                request = request,
                 credentialPresentationRequest = CredentialPresentationRequest.DCQLRequest(
                     buildDCQLQuery(
                         DCQLJsonClaimsQuery(
@@ -231,19 +239,19 @@ val AgentSdJwtTest by matrixSuite {
             val vp = presentationParameters.verifiablePresentations.values.flatten().firstOrNull()
                 .shouldBeInstanceOf<CreatePresentationResult.SdJwt>()
             // replace key binding of original vp.sdJwt (i.e. the part after the last `~`)
-            val freshKbJwt = createFreshSdJwtKeyBinding(it.challenge, it.verifierId)
+            val freshKbJwt = createFreshSdJwtKeyBinding(request.nonce, request.audience)
             val malformedVpSdJwt = vp.serialized.replaceAfterLast("~", freshKbJwt.substringAfterLast("~"))
 
             shouldThrowAny {
                 it.verifier.verifyPresentationSdJwt(
                     SdJwtSigned.parseCatching(malformedVpSdJwt).getOrThrow(),
-                    it.challenge
                 ).getOrThrow()
             }
         }
 
         "dcql: wrong challenge in key binding jwt" {
-            val malformedChallenge = it.challenge.reversed()
+            val request = it.verifier.createPresentationRequest()
+            val malformedChallenge = request.nonce.reversed()
             val presentationParameters = it.holder.createDefaultPresentation(
                 request = PresentationRequestParameters(
                     nonce = malformedChallenge,
@@ -262,13 +270,14 @@ val AgentSdJwtTest by matrixSuite {
                 .shouldBeInstanceOf<CreatePresentationResult.SdJwt>()
 
             shouldThrowAny {
-                it.verifier.verifyPresentationSdJwt(vp.sdJwt, it.challenge).getOrThrow()
+                it.verifier.verifyPresentationSdJwt(vp.sdJwt).getOrThrow()
             }
         }
 
         "dcql: revoked sd jwt" {
+            val request = it.verifier.createPresentationRequest()
             val presentationParameters = it.holder.createDefaultPresentation(
-                request = PresentationRequestParameters(nonce = it.challenge, audience = it.verifierId),
+                request = request,
                 credentialPresentationRequest = CredentialPresentationRequest.DCQLRequest(
                     buildDCQLQuery(
                         DCQLJsonClaimsQuery(
@@ -289,7 +298,7 @@ val AgentSdJwtTest by matrixSuite {
                         storeEntry.sdJwt.statusElement.shouldBeInstanceOf<StatusListInfo>().index,
                     ) shouldBe true
                 }
-            it.verifier.verifyPresentationSdJwt(vp.sdJwt, it.challenge).getOrThrow()
+            it.verifier.verifyPresentationSdJwt(vp.sdJwt).getOrThrow()
                 .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
                 .freshnessSummary.tokenStatusValidationResult
                 .shouldBeInstanceOf<TokenStatusValidationResult.Invalid>()
@@ -306,15 +315,19 @@ val AgentSdJwtTest by matrixSuite {
                 verifyJwsObjectIntegrity = VerifyStatusListTokenHAIP(),
             )
 
-            val haipVerifier = VerifierAgent(
-                identifier = it.verifierId,
-                validatorSdJwt = ValidatorSdJwt(
-                    validator = Validator(tokenStatusResolver = haipTokenStatusResolver),
+            val haipVerifier = NonceChallengeVerifier(
+                verifierId = it.verifierId,
+                verifier = VerifierAgent(
+                    identifier = it.verifierId,
+                    validatorSdJwt = ValidatorSdJwt(
+                        validator = Validator(tokenStatusResolver = haipTokenStatusResolver),
+                    ),
                 ),
             )
+            val request = haipVerifier.createPresentationRequest()
 
             val presentationParameters = it.holder.createDefaultPresentation(
-                request = PresentationRequestParameters(nonce = it.challenge, audience = it.verifierId),
+                request = request,
                 credentialPresentationRequest = CredentialPresentationRequest.DCQLRequest(
                     buildDCQLQuery(
                         DCQLJsonClaimsQuery(
@@ -327,7 +340,7 @@ val AgentSdJwtTest by matrixSuite {
             val vp = presentationParameters.verifiablePresentations.values.first().first()
                 .shouldBeInstanceOf<CreatePresentationResult.SdJwt>()
 
-            haipVerifier.verifyPresentationSdJwt(vp.sdJwt, it.challenge).getOrThrow()
+            haipVerifier.verifyPresentationSdJwt(vp.sdJwt).getOrThrow()
                 .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
                 .freshnessSummary.tokenStatusValidationResult
                 .shouldBeInstanceOf<TokenStatusValidationResult.Valid>()
@@ -352,15 +365,19 @@ val AgentSdJwtTest by matrixSuite {
                 verifyJwsObjectIntegrity = VerifyStatusListTokenHAIP(),
             )
 
-            val haipVerifier = VerifierAgent(
-                identifier = it.verifierId,
-                validatorSdJwt = ValidatorSdJwt(
-                    validator = Validator(tokenStatusResolver = haipTokenStatusResolver),
+            val haipVerifier = NonceChallengeVerifier(
+                verifierId = it.verifierId,
+                verifier = VerifierAgent(
+                    identifier = it.verifierId,
+                    validatorSdJwt = ValidatorSdJwt(
+                        validator = Validator(tokenStatusResolver = haipTokenStatusResolver),
+                    ),
                 ),
             )
+            val request = haipVerifier.createPresentationRequest()
 
             val presentationParameters = it.holder.createDefaultPresentation(
-                request = PresentationRequestParameters(nonce = it.challenge, audience = it.verifierId),
+                request = request,
                 credentialPresentationRequest = CredentialPresentationRequest.DCQLRequest(
                     buildDCQLQuery(
                         DCQLJsonClaimsQuery(
@@ -373,7 +390,7 @@ val AgentSdJwtTest by matrixSuite {
             val vp = presentationParameters.verifiablePresentations.values.first().first()
                 .shouldBeInstanceOf<CreatePresentationResult.SdJwt>()
 
-            val test = haipVerifier.verifyPresentationSdJwt(vp.sdJwt, it.challenge).getOrThrow()
+            val test = haipVerifier.verifyPresentationSdJwt(vp.sdJwt).getOrThrow()
 
             test.shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
                 .freshnessSummary.tokenStatusValidationResult